マイページ・プロフィール設定修正

+++ src/database/migrations/2024_11_23_111116_create_clear_storage_files_table.php
@@ -13,15 +13,29 @@ return new class extends Migration
     public function up(): void
     {
         if (app()->environment('local')) {
-            Storage::disk('public')->deleteDirectory('items');
+
+            if (Storage::disk('public')->exists('items')) {
+                Storage::disk('public')->deleteDirectory('items');
+            }
+
+            if (Storage::disk('public')->exists('profile_images')) {
+                Storage::disk('public')->deleteDirectory('profile_images');
+            }
 
             Storage::disk('public')->makeDirectory('items');
+            Storage::disk('public')->makeDirectory('profile_images');
         }
 
 
         if (app()->environment('production')) {
 
-            Storage::disk('s3')->deleteDirectory('items');
+            if (Storage::disk('s3')->exists('items')) {
+                Storage::disk('s3')->deleteDirectory('items');
+            }
+
+            if (Storage::disk('s3')->exists('profile_images')) {
+                Storage::disk('s3')->deleteDirectory('profile_images');
+            }
         }
     }
